+++ frontend/ui/src/app/interfaces/models/history/event.ts
@@ -5,5 +5,6 @@ export interface Event {
   eventMonth: number;
   eventYear: number;
   description: string;
+  yearAbbreviation: string;
   timeline: string;
 }