+++ veltahleon/src/main/java/com/testament/veltahleon/dto/history/TimelineDTO.java
@@ -1,16 +1,20 @@
-package com.testament.veltahleon.dto.history;
-
-import lombok.*;
-
-import java.util.List;
-
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class TimelineDTO {
-    private Long id;
-    private String name;
-    private List<Long> events;
-}
+//package com.testament.veltahleon.dto.history;
+//
+//import lombok.*;
+//
+//import java.util.List;
+//
+//@Builder
+//@AllArgsConstructor
+//@NoArgsConstructor
+//@Getter
+//@Setter
+//public class TimelineDTO {
+//    private Long id;
+//    private String name;
+//    private Integer beginningYear;
+//    private String beginningYearAbbreviation;
+//    private Integer endingYear;
+//    private String endingYearAbbreviation;
+//    private List<Long> events;
+//}