+++ veltahleon/src/main/java/com/testament/veltahleon/rest/history/EventController.java
@@ -1,5 +1,7 @@
 package com.testament.veltahleon.rest.history;
 
+import com.testament.veltahleon.dto.history.EventDTO;
+import com.testament.veltahleon.mappers.history.EventMapper;
 import com.testament.veltahleon.model.entities.history.Event;
 import com.testament.veltahleon.responses.CustomResponse;
 import com.testament.veltahleon.services.ifc.history.EventService;
@@ -28,14 +30,18 @@ public class EventController {
     @Autowired
     private EventService eventService;
 
+    @Autowired
+    private EventMapper eventMapper;
+
     @GetMapping("/events")
     public ResponseEntity<CustomResponse> getPaginatedEvents(@RequestParam(value = "pageNumber", defaultValue = "0") int pageNumber, @RequestParam(value = "pageSize", defaultValue = "10") int pageSize) {
         List<Event> events = (List<Event>) eventService.getPaginatedEvents((pageNumber - 1), pageSize);
+        List<EventDTO> eventsDTO = events.stream().map(e -> eventMapper.convertToDTO(e)).toList();
         return ResponseEntity.ok(CustomResponse.builder()
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("dataRetrieved", events))
+                .data(Map.of("dataRetrieved", eventsDTO))
                 .message(events.size() + " events retrieved from page: " + pageNumber)
                 .build()
         );
@@ -44,12 +50,27 @@ public class EventController {
     @GetMapping("/events/all")
     public ResponseEntity<CustomResponse> getAllEvents() {
         List<Event> events = (List<Event>) eventService.getEvents();
+        List<EventDTO> eventsDTO = events.stream().map(e -> eventMapper.convertToDTO(e)).toList();
         return ResponseEntity.ok(CustomResponse.builder()
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("dataRetrieved", events))
+                .data(Map.of("dataRetrieved", eventsDTO))
                 .message("All events retrieved!")
+                .build()
+        );
+    }
+
+    @GetMapping("/events/{timeline}")
+    public ResponseEntity<CustomResponse> getEventsByTimelineName(@PathVariable(value = "timeline") String timeline) {
+        List<Event> events = (List<Event>) eventService.getEventsByTimeline(timeline);
+        List<EventDTO> eventsDTO = events.stream().map(e -> eventMapper.convertToDTO(e)).toList();
+        return ResponseEntity.ok(CustomResponse.builder()
+                .timestamp(LocalDateTime.now())
+                .status(HttpStatus.OK)
+                .statusCode(HttpStatus.OK.value())
+                .data(Map.of("dataRetrieved", eventsDTO))
+                .message(eventsDTO.size() + " events for " + timeline + " timeline retrieved!")
                 .build()
         );
     }
@@ -60,7 +81,7 @@ public class EventController {
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("datumRetrieved", eventService.getEventByID(id)))
+                .data(Map.of("datumRetrieved", eventMapper.convertToDTO(eventService.getEventByID(id))))
                 .message("Event retrieved!")
                 .build()
         );
@@ -68,11 +89,12 @@ public class EventController {
 
     @GetMapping("/event/eventYear")
     public ResponseEntity<CustomResponse> getEventsByEventYear(@RequestParam(value = "year") int year) {
+        List<EventDTO> eventsDTO = eventService.getEventsByEventYear(year).stream().map(e -> eventMapper.convertToDTO(e)).toList();
         return ResponseEntity.ok(CustomResponse.builder()
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("dataRetrieved", eventService.getEventsByEventYear(year)))
+                .data(Map.of("dataRetrieved", eventsDTO))
                 .message("Event retrieved!")
                 .build()
         );
@@ -80,23 +102,25 @@ public class EventController {
 
     @GetMapping("/event/{eventDay}/{eventMonth}/{eventYear}")
     public ResponseEntity<CustomResponse> getEventByEventDayAndEventMonthAndEventYear(@PathVariable(value = "eventDay") int eventDay, @PathVariable(value = "eventMonth") int eventMonth, @PathVariable(value = "eventYear") int eventYear) {
+        EventDTO eventDTO = eventMapper.convertToDTO(eventService.getEventByEventDayAndEventMonthAndEventYear(eventDay, eventMonth, eventYear));
         return ResponseEntity.ok(CustomResponse.builder()
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("datumRetrieved", eventService.getEventByEventDayAndEventMonthAndEventYear(eventDay, eventMonth, eventYear)))
+                .data(Map.of("datumRetrieved", eventDTO))
                 .message("Event retrieved!")
                 .build()
         );
     }
 
-    @PostMapping("/save/event")
-    public ResponseEntity<CustomResponse> saveEvent(@RequestBody @Valid Event event) {
+    @PostMapping("/save/event/{timeline}")
+    public ResponseEntity<CustomResponse> saveEvent(@RequestBody @Valid EventDTO eventDTO, @PathVariable(value = "timeline") String timeline) {
+        Event event = eventMapper.convertToEntity(eventDTO);
         return ResponseEntity.ok(CustomResponse.builder()
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("dataSaved", eventService.saveEvent(event)))
+                .data(Map.of("dataSaved", eventMapper.convertToDTO(eventService.saveEvent(event, timeline))))
                 .message("Event saved!")
                 .build()
         );
@@ -127,12 +151,13 @@ public class EventController {
     }
 
     @PutMapping("/modify/event/{id}")
-    public ResponseEntity<CustomResponse> modifyEvent(@PathVariable Long id, @RequestBody Event event) {
+    public ResponseEntity<CustomResponse> modifyEvent(@PathVariable Long id, @RequestBody EventDTO eventDTO) {
+        Event event = eventMapper.convertToEntity(eventDTO);
         return ResponseEntity.ok(CustomResponse.builder()
                 .timestamp(LocalDateTime.now())
                 .status(HttpStatus.OK)
                 .statusCode(HttpStatus.OK.value())
-                .data(Map.of("dataUpdated", eventService.modifyEvent(id, event)))
+                .data(Map.of("dataUpdated", eventMapper.convertToDTO(eventService.modifyEvent(id, event))))
                 .message("Event updated!")
                 .build()
         );